requirements.txt
python-telegram-bot
Pillow
piexif
ffmpeg-python
